Add email input to registration declaration page, use common page model and formprovider for all declaration pages, add declaration email feature toggle
