+++ app/forms/DeclarationFormProvider.scala
@@ -18,35 +18,46 @@ package forms
 
 import javax.inject.Inject
 import forms.mappings.Mappings
-import models.core.pages.FullName
+import models.core.pages.{Declaration, FullName}
 import play.api.data.Form
 import play.api.data.Forms.{mapping, optional}
 
 class DeclarationFormProvider @Inject() extends Mappings {
 
-  def apply(): Form[FullName] =
-  Form(
-    mapping(
-      "firstName" -> text("declaration.error.firstName.required")
-        .verifying(
+  def apply(): Form[Declaration] =
+    Form(
+      mapping(
+        "" -> fullName,
+        "email" -> optional(text().verifying(
           firstError(
-            maxLength(35, s"declaration.error.firstName.length"),
-            isNotEmpty("firstName", s"declaration.error.firstName.required"),
-            regexp(Validation.nameRegex, s"declaration.error.firstName.invalid")
-          )),
-      "middleName" -> optional(text()
-        .verifying(
-          firstError(
-            maxLength(35, s"declaration.error.middleName.length"),
-            regexp(Validation.nameRegex, s"declaration.error.middleName.invalid"))
+            maxLength(35, s"declaration.error.email.length"),
+            regexp(Validation.emailRegex, s"declaration.error.email.invalid"))
+        ))
+      )(Declaration.apply)(Declaration.unapply)
+    )
+
+  val fullName = mapping(
+
+    "firstName" -> text("declaration.error.firstName.required")
+      .verifying(
+        firstError(
+          maxLength(35, s"declaration.error.firstName.length"),
+          isNotEmpty("firstName", s"declaration.error.firstName.required"),
+          regexp(Validation.nameRegex, s"declaration.error.firstName.invalid")
         )),
-      "lastName" -> text("declaration.error.lastName.required")
-        .verifying(
-          firstError(
-            maxLength(35, s"declaration.error.lastName.length"),
-            isNotEmpty("lastName", s"declaration.error.lastName.required"),
-            regexp(Validation.nameRegex, s"declaration.error.lastName.invalid")
-          ))
-    )(FullName.apply)(FullName.unapply)
-  )
+    "middleName" -> optional(text()
+      .verifying(
+        firstError(
+          maxLength(35, s"declaration.error.middleName.length"),
+          regexp(Validation.nameRegex, s"declaration.error.middleName.invalid"))
+      )),
+    "lastName" -> text("declaration.error.lastName.required")
+      .verifying(
+        firstError(
+          maxLength(35, s"declaration.error.lastName.length"),
+          isNotEmpty("lastName", s"declaration.error.lastName.required"),
+          regexp(Validation.nameRegex, s"declaration.error.lastName.invalid")
+        ))
+  )(FullName.apply)(FullName.unapply)
+
 }
