+++ app/views/playback/DeclarationView.scala.html
@@ -29,7 +29,7 @@
 @(form: Form[_], affinityGroup : AffinityGroup, onSubmit: Call)(implicit request: Request[_], messages: Messages)
 
 @main_template(
-    title = s"${errorPrefix(form)} ${messages("declaration.changes.noChanges.title")}"
+    title = s"${errorPrefix(form)} ${messages("declaration.title")}"
     ) {
 
     @formHelper(action = onSubmit, 'autoComplete -> "off") {
@@ -38,40 +38,40 @@
 
         @components.error_summary(form.errors)
 
-        @components.heading("declaration.changes.noChanges.heading")
+        @components.heading("declaration.heading")
 
         @components.input_text(
             field = form("firstName"),
-            label = messages("declaration.changes.noChanges.firstName"),
+            label = messages("declaration.firstName"),
             ariaRequired = true
         )
 
         @components.input_text(
             field = form("middleName"),
-            label = messages("declaration.changes.noChanges.middleName")
+            label = messages("declaration.middleName")
         )
 
         @components.input_text(
             field = form("lastName"),
-            label = messages("declaration.changes.noChanges.lastName"),
+            label = messages("declaration.lastName"),
             ariaRequired = true
         )
 
 
-        @if(appConfig.variationsDeclarationEmailEnabled) {
-            <p>@messages("declaration.changes.noChanges.paragraph1")</p>
-            <p>@messages("declaration.changes.noChanges.paragraph2")</p>
+        @if(appConfig.declarationEmailEnabled) {
+            <p>@messages("declaration.paragraph1")</p>
+            <p>@messages("declaration.paragraph2")</p>
             @components.input_text(
                 field = form("email"),
                 inputClass = Some("declaration-email"),
-                label = messages("declaration.changes.noChanges.email")
+                label = messages("declaration.email")
             )
         }
 
         @if(affinityGroup == AffinityGroup.Agent) {
-            @components.warning("declaration.changes.noChanges.agent")
+            @components.warning("declaration.variation.agent")
         } else {
-            @components.warning("declaration.changes.noChanges.individual")
+            @components.warning("declaration.variation.individual")
         }
 
         @components.submit_button(Some("site.confirm-send"))
