+++ conf/application.conf
@@ -128,9 +128,7 @@ microservice {
           livingSettlorBusiness.enabled = false
         }
         variations = true
-        playback {
-            declaration.email.enabled = true
-        }
+        declaration.email.enabled = true
       }
 
       trusts {
