+++ conf/messages.en
@@ -928,6 +928,9 @@ declaration.heading = Declaration
 declaration.firstName = Your first name
 declaration.middleName = Your middle names (optional)
 declaration.lastName = Your last name
+declaration.paragraph1 = Please give us your email address to get a confirmation of your declaration.
+declaration.paragraph2 = If you do not enter your email address, you will not receive a confirmation email with the declaration reference number.
+declaration.email = Email address (optional)
 declaration.error.firstName.required = Enter your first name
 declaration.error.firstName.invalid = Your first name must only include letters a to z, numbers 0 to 9, ampersands (&), apostrophes, commas, forward slashes, full stops, hyphens, round brackets and spaces
 declaration.error.lastName.required = Enter your last name
@@ -936,29 +939,13 @@ declaration.error.firstName.length = Your first name must be 35 characters or le
 declaration.error.middleName.length = Your middle names must be 35 characters or less
 declaration.error.middleName.invalid = Your middle names must only include letters a to z, numbers 0 to 9, ampersands (&), apostrophes, commas, forward slashes, full stops, hyphens, round brackets and spaces
 declaration.error.lastName.length = Your last name must be 35 characters or less
+declaration.error.email.length = Your email must be 35 characters or less
+declaration.error.email.invalid = Enter a valid email address with no spaces, like name@example.com
 declaration.agent.warning = I confirm that my client has taken all reasonable steps to obtain up to date and accurate information for all of the entities given in this registration. I understand that if my client knowingly provides false information and cannot demonstrate that they have taken all reasonable steps, they could be subject to penalties.
 declaration.individual.warning = I confirm that I have taken all reasonable steps to obtain up to date and accurate information for all of the entities given in this registration. I understand that if I knowingly provide false information and I cannot demonstrate that I have taken all reasonable steps, I could be subject to penalties.
 
-declaration.changes.noChanges.title = Declaration
-declaration.changes.noChanges.heading = Declaration
-declaration.changes.noChanges.firstName = Your first name
-declaration.changes.noChanges.middleName = Your middle names (optional)
-declaration.changes.noChanges.lastName = Your last name
-declaration.changes.noChanges.paragraph1 = Please give us your email address to get a confirmation of your declaration.
-declaration.changes.noChanges.paragraph2 = If you do not enter your email address, you will not receive a confirmation email with the declaration reference number.
-declaration.changes.noChanges.email = Email address (optional)
-declaration.changes.noChanges.error.firstName.required = Enter your first name
-declaration.changes.noChanges.error.firstName.invalid = Your first name must only include letters a to z, numbers 0 to 9, ampersands (&), apostrophes, commas, forward slashes, full stops, hyphens, round brackets and spaces
-declaration.changes.noChanges.error.lastName.required = Enter your last name
-declaration.changes.noChanges.error.lastName.invalid = Your last name must only include letters a to z, numbers 0 to 9, ampersands (&), apostrophes, commas, forward slashes, full stops, hyphens, round brackets and spaces
-declaration.changes.noChanges.error.firstName.length = Your first name must be 35 characters or less
-declaration.changes.noChanges.error.middleName.length = Your middle names must be 35 characters or less
-declaration.changes.noChanges.error.middleName.invalid = Your middle names must only include letters a to z, numbers 0 to 9, ampersands (&), apostrophes, commas, forward slashes, full stops, hyphens, round brackets and spaces
-declaration.changes.noChanges.error.lastName.length = Your last name must be 35 characters or less
-declaration.changes.noChanges.error.email.length = Your email must be 35 characters or less
-declaration.changes.noChanges.error.email.invalid = Enter a valid email address with no spaces, like name@example.com
-declaration.changes.noChanges.agent.warning = I confirm that my client has taken all reasonable steps to obtain up to date and accurate information for all of the entities given in this registration. I understand that if my client knowingly provides false information and cannot demonstrate that they have taken all reasonable steps, they could be subject to penalties.
-declaration.changes.noChanges.individual.warning = I confirm that the information given is true and complete to the best of my knowledge and belief. I will ensure it is regularly kept up to date as required including changes of address. If I discover that I have made an error or something has changed, I will update this register.
+declaration.variation.agent.warning = I confirm that my client has taken all reasonable steps to obtain up to date and accurate information for all of the entities given in this registration. I understand that if my client knowingly provides false information and cannot demonstrate that they have taken all reasonable steps, they could be subject to penalties.
+declaration.variation.individual.warning = I confirm that the information given is true and complete to the best of my knowledge and belief. I will ensure it is regularly kept up to date as required including changes of address. If I discover that I have made an error or something has changed, I will update this register.
 
 summaryAnswerPage.title = Draft copy of the trust’s registration
 summaryAnswerPage.heading = Draft copy of the trust’s registration
