--- test/forms/playback/DeclarationFormProviderSpec.scala
@@ -1,135 +1,0 @@
-/*
- * Copyright 2019 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package forms.playback
-
-import forms.Validation
-import forms.behaviours.StringFieldBehaviours
-import play.api.data.FormError
-import wolfendale.scalacheck.regexp.RegexpGen
-
-class DeclarationFormProviderSpec extends StringFieldBehaviours {
-
-  val form = new DeclarationFormProvider()()
-
-  ".firstName" must {
-
-    val fieldName = "firstName"
-    val requiredKey = "declaration.changes.noChanges.error.firstName.required"
-    val lengthKey = "declaration.changes.noChanges.error.firstName.length"
-    val maxLength = 35
-
-    behave like fieldThatBindsValidData(
-      form,
-      fieldName,
-      RegexpGen.from(Validation.nameRegex)
-    )
-
-    behave like fieldWithMaxLength(
-      form,
-      fieldName,
-      maxLength = maxLength,
-      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
-    )
-
-    behave like mandatoryField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey)
-    )
-
-    behave like nonEmptyField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey, Seq(fieldName))
-    )
-  }
-
-  ".middleName" must {
-
-    val fieldName = "middleName"
-    val lengthKey = "declaration.changes.noChanges.error.middleName.length"
-    val maxLength = 35
-
-    behave like fieldWithMaxLength(
-      form,
-      fieldName,
-      maxLength = maxLength,
-      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
-    )
-
-    behave like optionalField(
-      form,
-      fieldName,
-      validDataGenerator = RegexpGen.from(Validation.nameRegex)
-    )
-  }
-
-  ".lastName" must {
-
-    val fieldName = "lastName"
-    val requiredKey = "declaration.changes.noChanges.error.lastName.required"
-    val lengthKey = "declaration.changes.noChanges.error.lastName.length"
-    val maxLength = 35
-
-    behave like fieldThatBindsValidData(
-      form,
-      fieldName,
-      RegexpGen.from(Validation.nameRegex)
-    )
-
-    behave like fieldWithMaxLength(
-      form,
-      fieldName,
-      maxLength = maxLength,
-      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
-    )
-
-    behave like mandatoryField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey)
-    )
-
-    behave like nonEmptyField(
-      form,
-      fieldName,
-      requiredError = FormError(fieldName, requiredKey, Seq(fieldName))
-    )
-  }
-
-  ".email" must {
-
-    val fieldName = "email"
-    val lengthKey = "declaration.changes.noChanges.error.email.length"
-    val maxLength = 35
-
-    behave like fieldWithMaxLength(
-      form,
-      fieldName,
-      maxLength = maxLength,
-      lengthError = FormError(fieldName, lengthKey, Seq(maxLength))
-    )
-
-    behave like optionalField(
-      form,
-      fieldName,
-      validDataGenerator = RegexpGen.from(Validation.emailRegex)
-    )
-  }
-
-
-}
