+++ test/generators/ModelGenerators.scala
@@ -18,8 +18,8 @@ package generators
 
 import java.time.LocalDate
 
-import models.core.pages.{FullName, IndividualOrBusiness, InternationalAddress, UKAddress}
-import models.playback.pages.{Declaration, DeclarationWhatNext}
+import models.core.pages.{Declaration, FullName, IndividualOrBusiness, InternationalAddress, UKAddress}
+import models.playback.pages.DeclarationWhatNext
 import models.registration.pages._
 import org.scalacheck.Arbitrary.arbitrary
 import org.scalacheck.{Arbitrary, Gen}
@@ -80,7 +80,7 @@ trait ModelGenerators {
     }
   }
 
-  implicit lazy val arbitraryDeclarationChangesNoChanges : Arbitrary[models.playback.pages.Declaration] = {
+  implicit lazy val arbitraryDeclaration : Arbitrary[Declaration] = {
     Arbitrary {
       for {
         str <- arbitrary[String]
