+++ test/generators/PageGenerators.scala
@@ -157,9 +157,6 @@ trait PageGenerators {
   implicit lazy val arbitraryDeclarationPage: Arbitrary[pages.DeclarationPage.type] =
     Arbitrary(pages.DeclarationPage)
 
-  implicit lazy val arbitraryDeclarationChangesNoChangesPage: Arbitrary[pages.playback.DeclarationPage.type] =
-    Arbitrary(pages.playback.DeclarationPage)
-
   implicit lazy val arbitraryWhatTypeOfBeneficiaryPage: Arbitrary[WhatTypeOfBeneficiaryPage.type] =
     Arbitrary(WhatTypeOfBeneficiaryPage)
