+++ test/generators/UserAnswersGenerator.scala
@@ -75,7 +75,6 @@ trait UserAnswersGenerator extends TryValues {
     arbitrary[(ShareClassPage, JsValue)] ::
     arbitrary[(PropertyOrLandUKAddressPage, JsValue)] ::
     arbitrary[(pages.DeclarationPage.type, JsValue)] ::
-    arbitrary[(pages.playback.DeclarationPage.type, JsValue)] ::
     arbitrary[(WhatTypeOfBeneficiaryPage.type, JsValue)] ::
     arbitrary[(AgentInternationalAddressPage.type, JsValue)] ::
     arbitrary[(AgentUKAddressPage.type, JsValue)] ::
